Data_Engineering_Customers_fans_Pipeline/requirements.txt
psycopg2-binary
requests
python-dateutil
phonenumbers
fastapi
uvicorn
pydantic
